# Scripts: Optimize default Webpack configuration