--- a/packages/scripts/config/webpack.config.js
+++ b/packages/scripts/config/webpack.config.js
@@ -97,10 +97,13 @@ const config = {
 			{
 				test: /\.js$/,
 				exclude: /node_modules/,
-				use: {
-					loader: require.resolve( 'babel-loader' ),
-					options: getBabelLoaderOptions(),
-				},
+				use: [
+					'thread-loader',
+					{
+						loader: require.resolve( 'babel-loader' ),
+						options: getBabelLoaderOptions(),
+					},
+				],
 			},
 		],
 	},