--- a/packages/scripts/package.json
+++ b/packages/scripts/package.json
@@ -50,6 +50,7 @@
 		"source-map-loader": "^0.2.4",
 		"stylelint": "^9.10.1",
 		"stylelint-config-wordpress": "^13.1.0",
+		"thread-loader": "^2.1.2",
 		"webpack": "4.8.3",
 		"webpack-bundle-analyzer": "^3.0.3",
 		"webpack-cli": "^3.1.2",