@nosolosw 2019-04-12
This is the result of my testing:

| | first time | second & subsequent times |
--- | --- | ---
| master | ~35s | ~20s
| PR | ~30s | ~10s

I've also played with what Ben suggested (a static number of workers from 1 to 4 + a dynamic number as in calypso): it didn't render any improvement over the PR here.

Interestingly, this is how my cores behave (tested it several times and can repro consistently, in an Intel® Core™ i7-5600U CPU @ 2.60GHz × 4). In `master` all my cores reach the 100% at some point but they require a bit more time to be coordinated:

![Screenshot from 2019-04-12 12-57-08-master](https://user-images.githubusercontent.com/583546/56032731-222b1c80-5d23-11e9-9e50-d9d47038a78b.png)

In this PR, they seem to work in a more coordinated fashion:

![Screenshot from 2019-04-12 12-54-17-withthread](https://user-images.githubusercontent.com/583546/56032755-2e16de80-5d23-11e9-912d-3ed71be4d3f3.png)