@aduth 2019-04-26
CHANGELOG notes should have been included here. I committed them directly to master in c8695f72db8dc5ecc00f70d6fe556dc63da1fd36 .